# docs/api/swagger.yaml
definitions:
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion:
    properties:
      amount:
        type: number
      created:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      title:
        type: string
      type:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.PromotionType'
      updated:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.PromotionType:
    enum:
    - regular
    - welcome_bonus
    type: string
    x-enum-varnames:
    - Regular
    - WelcomeBonus
...
info:
  contact: {}
paths:
...
  /api/v1/promotions:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all promotions
      produces:
      - application/json
      responses:
        "200":
          description: List of promotions
          schema:
            items:
              $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Get all promotions
      tags:
      - Promotions
    post:
      consumes:
      - application/json
      description: Create a new promotion with the provided details
      parameters:
      - description: Promotion details
        in: body
        name: promotion
        required: true
        schema:
          $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
      produces:
      - application/json
      responses:
        "200":
          description: Created promotion
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Create a new promotion
      tags:
      - Promotions
  /api/v1/promotions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a promotion using its unique ID
      parameters:
      - description: Promotion ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "404":
          description: Promotion not found
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Delete a promotion
      tags:
      - Promotions
    get:
      consumes:
      - application/json
      description: Retrieve a promotion using its unique ID
      parameters:
      - description: Promotion ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved promotion
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "404":
          description: Promotion not found
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Get a promotion by ID
      tags:
      - Promotions
    put:
      consumes:
      - application/json
      description: Update an existing promotion with the provided details
      parameters:
      - description: Updated promotion details
        in: body
        name: promotion
        required: true
        schema:
          $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
      produces:
      - application/json
      responses:
        "200":
          description: Updated promotion
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Update a promotion
      tags:
      - Promotions
...
swagger: "2.0"